requirements.txt
pillow==10.1.0
pyheif==0.7.1 